+++ Baselet/ant.xml
@@ -23,13 +23,7 @@
 	<property name="dir.export.plugin" location="${dir.export}/plugins" />
 	<property name="elements.src" location="../BaseletElements/src" />
 	<property name="elements.lib" location="../BaseletElements/lib" />
-
-	<property name="javacc.grammar.folder" location="${elements.src}/com/baselet/element/facet/common/CustomDrawingParser/" />
-	<property name="javacc.grammar.file" value="CustomDrawingParser.jj" />
-	<property name="javacc.grammar.path" location="${javacc.grammar.folder}/${javacc.grammar.file}" />
-	<!-- This folder must contain the javacc.jar minimum  6.1 for GWT compatibility (http://consoliii.blogspot.co.at/2014/04/creating-gwt-compatible-parser-using.html)-->
-	<!-- JavaCC can be obtained here https://java.net/projects/javacc/downloads -->
-	<property name="javacc.home" location="../javacc_distribution" />
+	<property name="elements.dir" location="../BaseletElements/" />
 
 	<!-- Classpaths -->
 	<path id="project.classpath.build">
@@ -49,6 +43,11 @@
 	<path id="project.classpath.lib">
 		<fileset dir="${dir.lib}" includes="*.jar" />
 	</path>
+	
+	<!-- Generates the parser from the grammar -->
+	<target name="gen_custom_drawing_parser" >
+		<ant dir="${elements.dir}" target="javacc_custdraw" />
+	</target>
 
 	<!-- Update MANIFEST.MF and includes new version and libs -->
 	<target name="update_manifest">
@@ -96,36 +95,9 @@
 			<replacefilter token="@release.year@" value="${release.year}" />
 		</replace>
 	</target>
-
-	<!-- JavaCC and JJDoc Targets -->
-	<target name="javacc_custdraw" description="Generates the java classes from the grammar file">
-		<javacc target="${javacc.grammar.path}" javacchome="${javacc.home}" />
-		<!-- after generation of the java classes refresh the folder so that eclipse recognizes the changes  -->
-		<eclipse.refreshLocal resource="${javacc.grammar.folder}" depth="infinite" />
-	</target>
-	
-	<target name="jjdoc_custdraw" description="Generates a simple BNF documentation for the grammar file">
-		<jjdoc target="${javacc.grammar.path}" javacchome="${javacc.home}" outputfile="${javacc.grammar.folder}/CustomDrawingGrammar.html" />
-		<!-- after generation of the java classes refresh the folder so that eclipse recognizes the changes  -->
-		<eclipse.refreshLocal resource="${javacc.grammar.folder}" depth="infinite" />
-	</target>
-	
-	<target name="javacc_jjdoc_clean" description="Deletes the files which are generated by javac_custdraw and jjdoc_custdraw.">
-		<delete file="${javacc.grammar.folder}/CustomDrawingParser.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/CustomDrawingParserConstants.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/CustomDrawingParserTokenManager.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/JavaCharStream.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/ParseException.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/Provider.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/StreamProvider.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/StringProvider.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/Token.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/TokenMgrException.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/CustomDrawingGrammar.html" failonerror="false"/>
-	</target>
 	
 	<!-- Compile the project -->
-	<target name="javac" depends="update_manifest, update_html, javacc_custdraw">
+	<target name="javac" depends="update_manifest, update_html, gen_custom_drawing_parser">
 		<mkdir dir="${dir.bin}" />
 		<!-- Compile elements project (into the same outputdir) -->
 		<javac source="${javac.version}" target="${javac.version}" srcdir="${elements.src}" destdir="${dir.bin}" nowarn="true" includeantruntime="false">
