+++ BaseletElements/src/com/baselet/element/elementnew/uml/Class.java
@@ -7,8 +7,8 @@ import com.baselet.element.NewGridElement;
 import com.baselet.element.facet.Facet;
 import com.baselet.element.facet.PropertiesParserState;
 import com.baselet.element.facet.Settings;
-import com.baselet.element.facet.common.CustomDrawingFacet;
 import com.baselet.element.facet.common.SeparatorLineWithHalignChangeFacet;
+import com.baselet.element.facet.customdrawings.CustomDrawingFacet;
 import com.baselet.element.facet.specific.ActiveClassFacet;
 import com.baselet.element.facet.specific.InnerClassFacet;
 import com.baselet.element.facet.specific.TemplateClassFacet;
