+++ BaseletGWT/build.xml
@@ -11,12 +11,7 @@
 	<property name="gwt.module.name" value="com.baselet.gwt.BaseletGWT" />
 	<property name="gwt.sdk.location" value="C:/Program Files/Eclipse/plugins/com.google.gwt.eclipse.sdkbundle_2.7.0/gwt-2.7.0" />
 	<property name="elements.src.dir" location="../BaseletElements/src" />
-	<property name="javacc.grammar.folder" location="${elements.src.dir}/com/baselet/element/facet/common/CustomDrawingParser/" />
-	<property name="javacc.grammar.file" value="CustomDrawingParser.jj" />
-	<property name="javacc.grammar.path" location="${javacc.grammar.folder}/${javacc.grammar.file}" />
-	<!-- This folder must contain the javacc.jar minimum  6.1 for GWT compatibility (http://consoliii.blogspot.co.at/2014/04/creating-gwt-compatible-parser-using.html)-->
-	<!-- JavaCC can be obtained here https://java.net/projects/javacc/downloads -->
-	<property name="javacc.home" location="../javacc_distribution" />
+	<property name="elements.dir" location="../BaseletElements/" />
 
 	<path id="project.classpath">
 		<fileset dir="${lib.dir}">
@@ -26,32 +21,15 @@
 		<pathelement location="${gwt.sdk.location}/gwt-dev.jar" />
 		<pathelement location="${elements.src.dir}" />
 	</path>
+	
 
-	<!-- JavaCC and JJDoc Targets -->
-	<target name="javacc_custdraw" description="Generates the java classes from the grammar file">
-		<javacc target="${javacc.grammar.path}" javacchome="${javacc.home}" />
+	<!-- Generates the parser from the grammar -->
+	<target name="gen_custom_drawing_parser" >
+		<ant dir="${elements.dir}" target="javacc_custdraw" />
 	</target>
-	
-	<target name="jjdoc_custdraw" description="Generates a simple BNF documentation for the grammar file">
-		<jjdoc target="${javacc.grammar.path}" javacchome="${javacc.home}" outputfile="${javacc.grammar.folder}/CustomDrawingGrammar.html" />
-	</target>
-	
-	<target name="javacc_jjdoc_clean" description="Deletes the files which are generated by javac_custdraw and jjdoc_custdraw.">
-		<delete file="${javacc.grammar.folder}/CustomDrawingParser.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/CustomDrawingParserConstants.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/CustomDrawingParserTokenManager.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/JavaCharStream.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/ParseException.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/Provider.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/StreamProvider.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/StringProvider.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/Token.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/TokenMgrException.java" failonerror="false"/>
-		<delete file="${javacc.grammar.folder}/CustomDrawingGrammar.html" failonerror="false"/>
-	</target>
-	
+
 	<!-- GWT targets -->
-	<target name="gwt-compile-helper" depends="javacc_custdraw">
+	<target name="gwt-compile-helper" depends="gen_custom_drawing_parser">
 		<delete dir="${build.dir}" />
 		<java failonerror="true" fork="true" classname="com.google.gwt.dev.Compiler">
 			<classpath>
